requirements.txt
langchain
langchain-community
langgraph
langchain-groq
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
tiktoken